+++ src/Exception/InvalidAccessTokenException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Drip\Exception;
+
+class InvalidAccessTokenException extends DripException
+{
+}