+++ src/Exception/InvalidApiKeyException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Drip\Exception;
+
+class InvalidApiKeyException extends InvalidApiTokenException
+{
+}